bot/requirements.txt
aiogram==3.13.1
groq==0.11.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
aiosqlite==0.20.0
APScheduler==3.10.4
pdfplumber==0.11.4
python-dotenv==1.0.1
